window manager with tests
